moved userdb out to figure out error. At the moment, the server side is debugged, aside when trying to load the website from both ends. But at the moment, the run build and run dev actually work.
