+++ how-not-to-die/Server/index.ts
@@ -1,8 +1,8 @@
 import express from 'express';
 import dotenv from 'dotenv';
-import authRoutes from './src/routes/auth-routes';
-import userRoutes from './src/routes//api/user-routes';
-import sequelize from './src/config/connection';
+import authRoutes from './src/routes/auth-routes.js';
+import userRoutes from './src/routes//api/user-routes.js';
+import sequelize from './src/config/connection.js';
 
 dotenv.config();
