+++ how-not-to-die/Server/server.js
@@ -49,7 +49,7 @@ sequelize.authenticate()
     .then(() => console.log('Connected to PostgreSQL via Sequelize'))
     .catch((err) => console.error('Unable to connect to database:', err));
 // API route to fetch planets
-app.get('/api/planets', async (req, res) => {
+app.get('/api/planets', async (_req, res) => {
     try {
         const planets = await Planet.findAll();
         res.json(planets);
