+++ how-not-to-die/Server/src/middleware/auth.js
@@ -1,7 +1,27 @@
 import jwt from 'jsonwebtoken';
+import dotenv from 'dotenv';
+dotenv.config();
+const JWT_SECRET_KEY = process.env.JWT_SECRET_KEY;
+if (!JWT_SECRET_KEY) {
+    throw new Error('JWT_SECRET_KEY is not defined in environment variables.');
+}
+// Generate a token
 export const generateToken = (userId) => {
-    return jwt.sign({ id: userId }, process.env.JWT_SECRET_KEY, { expiresIn: '1h' });
+    return jwt.sign({ id: userId }, JWT_SECRET_KEY, { expiresIn: '1h' });
 };
-export const verifyToken = (token) => {
-    return jwt.verify(token, process.env.JWT_SECRET_KEY);
+// ✅ Correct Type Definition for Middleware
+export const verifyToken = (req, res, next) => {
+    const token = req.header('Authorization')?.split(' ')[1];
+    if (!token) {
+        res.status(401).json({ error: 'Access denied. No token provided.' });
+        return;
+    }
+    try {
+        const decoded = jwt.verify(token, JWT_SECRET_KEY);
+        req.user = decoded; // Attach user info to request object
+        next();
+    }
+    catch (error) {
+        res.status(403).json({ error: 'Invalid or expired token.' });
+    }
 };
