+++ how-not-to-die/Server/src/middleware/auth.ts
@@ -1,15 +1,17 @@
 import jwt from 'jsonwebtoken';
 import { Request, Response, NextFunction } from 'express';
+import dotenv from 'dotenv';
+dotenv.config();
 
-const JWT_SECRET = process.env.JWT_SECRET_KEY as string;
+const JWT_SECRET_KEY = process.env.JWT_SECRET_KEY as string;
 
-if (!JWT_SECRET) {
+if (!JWT_SECRET_KEY) {
   throw new Error('JWT_SECRET_KEY is not defined in environment variables.');
 }
 
 // Generate a token
 export const generateToken = (userId: number): string => {
-  return jwt.sign({ id: userId }, JWT_SECRET, { expiresIn: '1h' });
+  return jwt.sign({ id: userId }, JWT_SECRET_KEY, { expiresIn: '1h' });
 };
 
 // ✅ Correct Type Definition for Middleware
@@ -22,7 +24,7 @@ export const verifyToken = (req: Request, res: Response, next: NextFunction): vo
   }
 
   try {
-    const decoded = jwt.verify(token, JWT_SECRET) as { id: number };
+    const decoded = jwt.verify(token, JWT_SECRET_KEY) as { id: number };
     (req as any).user = decoded; // Attach user info to request object
     next();
   } catch (error) {
