+++ how-not-to-die/Server/src/routes/api/user-routes.js
@@ -1,6 +1,6 @@
 import express from 'express';
-import User from './user';
-import { verifyToken } from './auth';
+import User from '../../models/user.js';
+import { verifyToken } from '../../middleware/auth.js';
 const router = express.Router();
 router.get('/', verifyToken, async (_req, res) => {
     const users = await User.findAll();
