+++ how-not-to-die/Server/src/routes/index.ts
@@ -1,5 +1,5 @@
-import authRoutes from './auth-routes';
-import userRoutes from '../routes/api/user-routes';
+import authRoutes from './auth-routes.js';
+import userRoutes from '../routes/api/user-routes.js';
