+++ how-not-to-die/Server/src/seeds/user-seeds.ts
@@ -1,4 +1,4 @@
-import User from '../models/user';
+import User from '../models/user.js';
 
 const seedUsers = async () => {
   await User.bulkCreate([
